tests/Aban_tests.py
# https://forum.learncodethehardway.com/t/testing-input-and-print/1757/10

from nose.tools import *
from unittest import mock
# from the module folder (Forest_Game) import the file Aban.py
# and then Abandoned is the class I'm testing!
import sys
sys.path.insert(0, '/Users/HomeFolder/projects/Forest_Game/Forest_Game')
from Aban import Abandoned
from charac_fg import Character_

# With NOSE this line isn't necessary, something like this IS necessary if you're just using unitttest
# class TestClass(unittest.TestCase):

# the setup and teardown methods always come before the methods to be tested (i think)
def setup():
    print("SETUP!")

def teardown():
    print("TEAR DOWN!")

def test_basic():
    print("I RAN!", end='')

def test_abandoned():
    aban = Abandoned()

    # I'm commenting out the .examine check for now b/c it's saying the characters don't
    # match even though I'm copying and pasting the text...

    # assert_equal(aban.examine,"""
    # ... it appears someone has been here recently. You see
    # an abandoned fire pit, some sticks lying next to the fire pit,
    # and you see a small green bag off next to some rocks.
    # """)
    assert_equal(aban.name, "Abandoned Fire Pit")
    assert_equal(aban.down, "-")


def test_pickupsticks():
    # stick = Abandoned()
    Character_.inven.append('sticks')
    Character_.get_inven()

    assert_equal(Character_.inven, ['sticks'])

def test_inspectbag():
    Character_.inven.append('steel')
    Character_.get_inven()

    # I'm not sure if this will work, b/c I'm not sure if sticks will also be there!
    assert_equal(Character_.inven, ['sticks','steel'])

def test_pickuprocks_maketorch():
    Character_.inven.append('flint')
    Character_.get_inven()

    assert_equal(Character_.inven, ['Torch!'])

# def test_maketorch():
#     # Character_.inven.append('sticks','steel','flint')
#     if 'sticks' and 'steel' and 'flint' in Character_.inven:
#         print("You've collected the items to create a Torch!")
#         Character_.inven.append('Torch!')
#         Character_.inven.remove('sticks','steel','flint')
#         Character_.get_inven()
#
#     assert_equal(Character_.inven, ['Torch!'])

# ok based upon this link: https://forum.learncodethehardway.com/t/testing-input-and-print/1757/10
# I think I need to create separate test functions for each possible
# input and I need to use a module called mock from unittest to create fake inputs...

def test_move_right():
    right = Abandoned()

    org_input = mock.builtins.input
    mock.builtins.input = lambda _: "right"
    # if anything causes an error it will most likely be this line...
    assert_equal(right.move(), 'river')
    mock.builtins.input = org_input



# def test_move():
#
#     ### LOOK UP HOW TO MAKE TESTS FOR METHODS THAT HAVE USER INPUT!!!!
#     self.input = input("Which direction would you like to go? \n")
#     super(Abandoned, self).move()
#     if self.input.lower() in ['right', 'east']:
#         self.right = 'river'
#         return self.right
#     elif self.input.lower() in ['left','west']:
#         self.nothing_()
#     elif self.input.lower() in ['down','south']:
#         self.down = 'hole'
#         return self.down
#     elif self.input.lower() in ['up','north']:
#         self.up = 'edge'
#         return self.up
#
#     right = Abandoned()
#     left = Abandoned()
#     up = Abandoned()
#     down = Abandoned()
#
#     right.move()
#
#     assert_equal(start)
